# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Reverb API
  version: 0.1.0
  description: >
    Reverb is a self-hosted music app unifying a Subsonic/Navidrome library,
    Spotify search, and spotDL one-click download. All paths are served under
    /api/v1. The local browser UI is implicitly authenticated as the single
    household owner ("local"); paired devices authenticate to /sync with Bearer
    sync tokens (issued from one-time pairing codes) and to P2P via libp2p peer
    IDs bound at pairing.
servers:
  - url: /api/v1
paths:
...
  /playlists:
    get:
      summary: List managed playlists
      responses:
        "200":
          description: every managed playlist in the household library
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SyncedPlaylist' } } } }
        "503": { description: playlist sync unavailable }
...
  /pairing/code:
    post:
      summary: Generate a single-use pairing code (manage-library capability required)
      responses:
        "200":
          description: pairing code
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string, example: "AB12-CD34" }
                  expiresAt: { type: integer, format: int64 }
                  qrPayload:
                    type: string
                    description: >
                      The code as a versioned pairing link for a QR code:
                      reverb://pair?v=1 with the code, its expiry (exp, Unix
                      seconds), this device's peer ID (peer) and each address
                      another device can dial it on (addr, repeated; LAN and
                      VPN). Redeem it with POST /p2p/pair/redeem-qr. Omitted
                      when this device has no dialable address.
                    example: "reverb://pair?v=1&code=AB12CD34&exp=1800000000&peer=12D3KooW...&addr=%2Fip4%2F192.168.1.20%2Ftcp%2F4331"
                  qrSvg:
                    type: string
                    description: qrPayload drawn as a QR code, as a standalone SVG document.
                required: [code, expiresAt]
        "503": { description: pairing unavailable }
...
  /p2p/peers:
    get:
      summary: List connected libp2p peers (manage-library capability required)
      description: >
        Lists transport-level connections. Discovery is not trust: mDNS and DHT
        connect to unpaired peers, which are refused by the sync and file
        protocols. Pairing state is not reflected here.
      responses:
        "200":
          description: connected peers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    peerId: { type: string }
                    addrs: { type: array, items: { type: string } }
                    conns: { type: integer }
                    connected: { type: boolean }
                  required: [peerId, addrs, conns, connected]
        "503": { description: p2p unavailable }
...
components:
...
  schemas:
...
    SyncedPlaylist:
      type: object
      required: [id, source, externalId, name, mode, syncEnabled, syncIntervalSec, autoDownload, lastSyncedAt, trackCount]
      properties:
        id: { type: string }
        source: { type: string, description: "local for a playlist made in Reverb." }
        externalId: { type: string }
        name: { type: string }
        coverUrl: { type: string }
        mode: { type: string, description: "once for a playlist whose tracklist replicates; synced for a mirror of an upstream playlist." }
        syncEnabled: { type: boolean }
        syncIntervalSec: { type: integer }
        autoDownload: { type: boolean }
        lastSyncedAt: { type: integer, format: int64 }
        trackCount: { type: integer }
...